pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.5",
    "faiss-cpu>=1.13.2",
    "hf-xet>=1.2.0",
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.8",
    "langchain>=0.2.0,<0.3.0",
    "langchain-community>=0.2.0,<0.3.0",
    "langchain-core>=0.2.36,<0.3.0",
    "langchain-ollama==0.1.3",
    "pdf2image>=1.17.0",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.19",
    "streamlit>=1.40.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
    "transformers>=4.57.3",
    "unstructured[all-docs]>=0.18.21",
    "uvicorn[standard]>=0.30.6",
]
[project.optional-dependencies]
vector-faiss = ["faiss-cpu"]
